@@ -20,7 +20,7 @@
 
     <properties>
         <revision>1.0.0</revision>
-        <java.version>1.8</java.version>
+        <java.version>17</java.version>
         <spring.boot.version>2.7.8</spring.boot.version>
         <druid.version>1.2.16</druid.version>
         <mysql.version>8.0.22</mysql.version>
@@ -37,7 +37,7 @@
         <hutool.version>5.8.25</hutool.version>
         <javase.verison>3.2.1</javase.verison>
         <core.verison>3.3.3</core.verison>
-        <sa.token.verison>1.34.0</sa.token.verison>
+        <sa.token.verison>1.37.0</sa.token.verison>
         <netty.verison>4.1.86.Final</netty.verison>
     </properties>
 
@@ -186,10 +186,9 @@
                 <artifactId>sa-token-jwt</artifactId>
                 <version>${sa.token.verison}</version>
             </dependency>
-
             <dependency>
                 <groupId>cn.dev33</groupId>
-                <artifactId>sa-token-dao-redis-jackson</artifactId>
+                <artifactId>sa-token-redis-jackson</artifactId>
                 <version>${sa.token.verison}</version>
             </dependency>
 
@@ -210,8 +209,8 @@
                     <version>3.8.1</version>
                     <configuration>
                         <skip>true</skip>
-                        <source>1.8</source>
-                        <target>1.8</target>
+                        <source>${java.version}</source>
+                        <target>${java.version}</target>
                         <annotationProcessorPaths>
                             <path>
                                 <groupId>org.projectlombok</groupId>
