@@ -59,9 +59,8 @@ public class WebUtil {
      * 针对集合 获取类
      **/
     public static Field getField(Object target, String property) {
-        if (target instanceof List<?>) {
-            List<?> targets = (List<?>) target;
-            return com.fast.core.common.util.RUtil.getField(targets.get(0).getClass(), property);
+        if (target instanceof List<?> targets) {
+            return RUtil.getField(targets.get(0).getClass(), property);
         }
         return RUtil.getField(target.getClass(), property);
     }
@@ -77,46 +76,46 @@ public class WebUtil {
         if (request == null) {
             return "0.0.0.0";
         }
-        String Xip = request.getHeader("X-Real-IP");
-        String XFor = request.getHeader("X-Forwarded-For");
-        String UNKNOWN_IP = "unknown";
-        if (StringUtils.isNotEmpty(XFor) && !UNKNOWN_IP.equalsIgnoreCase(XFor)) {
+        String xip = request.getHeader("X-Real-IP");
+        String xFor = request.getHeader("X-Forwarded-For");
+        String unknown = "unknown";
+        if (StringUtils.isNotEmpty(xFor) && !unknown.equalsIgnoreCase(xFor)) {
             // 多次反向代理后会有多个ip值，第一个ip才是真实ip
-            int index = XFor.indexOf(",");
+            int index = xFor.indexOf(",");
             if (index != -1) {
-                return XFor.substring(0, index);
+                return xFor.substring(0, index);
             } else {
-                return XFor;
+                return xFor;
             }
         }
 
-        XFor = Xip;
+        xFor = xip;
 
-        if (StringUtils.isNotEmpty(XFor) && !UNKNOWN_IP.equalsIgnoreCase(XFor)) {
-            return XFor;
+        if (StringUtils.isNotEmpty(xFor) && !unknown.equalsIgnoreCase(xFor)) {
+            return xFor;
         }
 
-        if (StringUtils.isBlank(XFor) || UNKNOWN_IP.equalsIgnoreCase(XFor)) {
-            XFor = request.getHeader("Proxy-Client-IP");
+        if (StringUtils.isBlank(xFor) || unknown.equalsIgnoreCase(xFor)) {
+            xFor = request.getHeader("Proxy-Client-IP");
         }
 
-        if (StringUtils.isBlank(XFor) || UNKNOWN_IP.equalsIgnoreCase(XFor)) {
-            XFor = request.getHeader("WL-Proxy-Client-IP");
+        if (StringUtils.isBlank(xFor) || unknown.equalsIgnoreCase(xFor)) {
+            xFor = request.getHeader("WL-Proxy-Client-IP");
         }
 
-        if (StringUtils.isBlank(XFor) || UNKNOWN_IP.equalsIgnoreCase(XFor)) {
-            XFor = request.getHeader("HTTP_CLIENT_IP");
+        if (StringUtils.isBlank(xFor) || unknown.equalsIgnoreCase(xFor)) {
+            xFor = request.getHeader("HTTP_CLIENT_IP");
         }
 
-        if (StringUtils.isBlank(XFor) || UNKNOWN_IP.equalsIgnoreCase(XFor)) {
-            XFor = request.getHeader("HTTP_X_FORWARDED_FOR");
+        if (StringUtils.isBlank(xFor) || unknown.equalsIgnoreCase(xFor)) {
+            xFor = request.getHeader("HTTP_X_FORWARDED_FOR");
         }
 
-        if (StringUtils.isBlank(XFor) || UNKNOWN_IP.equalsIgnoreCase(XFor)) {
-            XFor = request.getRemoteAddr();
+        if (StringUtils.isBlank(xFor) || unknown.equalsIgnoreCase(xFor)) {
+            xFor = request.getRemoteAddr();
         }
 
-        return XFor;
+        return xFor;
 
     }
 }
