@@ -10,7 +10,10 @@ import lombok.Data;
  **/
 @Data
 public class Query {
-    /** 当前记录起始索引 */
+    private String id;
+    /**
+     * 当前记录起始索引
+     */
     private Integer pageNum;
     /** 每页显示记录数 */
     private Integer pageSize;
