@@ -1,7 +1,6 @@
 package com.fast.common.entity.sys;
 
 import com.baomidou.mybatisplus.annotation.TableField;
-import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
 import com.fast.core.common.validate.annotation.Display;
 import com.fast.core.mybatis.model.BaseVersionEntity;
@@ -26,12 +25,7 @@ import lombok.experimental.Accessors;
 @ToString(callSuper = true)
 @JsonInclude(value=JsonInclude.Include.NON_NULL)
 public class SysEncodingSet extends BaseVersionEntity {
-	/**
-	 * id
-	 */
-	@TableId
-	@Display("id")
-	private String id;
+
 
 	/**
 	 * 规则代码
